print reveerse

+++ 0x05-pointers_arrays_strings/4-print_rev.c
@@ -11,7 +11,7 @@
  */
 void print_rev(char *s)
 {
-	int l;
+	long int l;
 
 	l = 0;
